refactor: reduce duplicated code of glyph-picker
